Disable Del Curr User, Ffox Add Shift fix, GeoLoc range fix, Add Employee Validate Fix

+++ resources/views/company-settings/index.blade.php
@@ -506,7 +506,11 @@ console.log('sessionPrimary' + sessionPrimary);
 
                                 <td>
                                     <a href="/user/{{ $user->user_id }}/edit"><i class="fa fa-edit"></i></a>
-                                    <a href="/confirm-delete/{{$user->user_id}}/{{$url}}" style="color: #990000;"><i class="fa fa-trash-o icon-padding"></i></a>
+                                    @if(session('id') != $user->user_id)
+                                        <a href="/confirm-delete/{{$user->user_id}}/{{$url}}" style="color: #990000;"><i class="fa fa-trash-o icon-padding"></i></a>
+                                    @else
+                                        <i class="fa fa-trash-o icon-padding"></i>
+                                    @endif
                                 </td>
                             </tr>
                         @endforeach
